split compile,prove,verify
