+++ circuit_tutorials/halo2/axiom-v0.3.0/float_radius/zk_execute/.gitignore
@@ -1,0 +1,3 @@
+/target
+/debug
+/data/*
