+++ circuit_tutorials/halo2/axiom-v0.3.0/float_radius/zk_execute/Cargo.toml
@@ -1,14 +1,26 @@
 [package]
-name = "sindri-quickstart"
+name = "zk-execute"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[[bin]]
+name = "compile"
+
+[[bin]]
+name = "prove"
+
+[[bin]]
+name = "verify"
 
 [dependencies]
+base64 = "0.21.2"
+include_dir = "0.7.3"
 flate2 = "1.0.20"
 reqwest = { version = "0.11.20", features = ["multipart", "json"]}
 serde = { version = "1.0", features = ["derive"]}
 serde_json = "1.0"
 tar = "0.4.35"
 tokio = { version = "1.32", features = ["full"] }
+
+
+halo2-base = { git = "https://github.com/axiom-crypto/halo2-lib", tag = "v0.3.0", default-features = false, features = ["halo2-axiom"]  }
