--- circuit_tutorials/halo2/axiom-v0.3.0/float_radius/zk_execute/README.md
@@ -1,7 +1,0 @@
-The code in `src/main.rs` will create a circuit, upload, and initiate compile.  Then from that compiled circuit, it will initiate a proof. From this directory, you can call
-```
-cargo run
-```
-which builds the executable and runs it.
-
-DISCLAIMER: This code does minimal return code processing and could be improved to return proofs or keys to a user.
