+++ circuit_tutorials/halo2/axiom-v0.3.0/float_radius/zk_execute/src/bin/verify.rs
@@ -1,0 +1,32 @@
+use std::{
+    env,
+    fs,
+    io::{BufWriter,Read,Write}
+};
+use serde_json::Value;
+use halo2_base::halo2_proofs::halo2curves::bn256::Fr;
+use radius_circuit::gadgets::FixedPointChip;
+use base64::{engine::general_purpose, Engine as _};
+
+
+#[tokio::main]
+async fn main() {
+
+    let mut file = fs::File::open("./data/prove_out.json").unwrap();
+    let mut data = String::new();
+    file.read_to_string(&mut data).unwrap();
+    
+    // figure out best way to ref quantization gadget
+    //decode the instance from string to a field element
+    // let instance_str = "MNMOJwYAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=";
+    // let field_instance = Fr::from_bytes(&general_purpose::STANDARD.decode(instance_str).unwrap().try_into().unwrap()).unwrap();
+
+    // //instantiate the Fixed Point Chip (which will dequantize the instance variable)
+    // let lookup_bits = 12;
+    // const PRECISION_BITS: u32 = 32;
+    // let fixed_point_chip = FixedPointChip::<Fr, PRECISION_BITS>::default(lookup_bits);
+
+    // let radius = fixed_point_chip.dequantization(field_instance);
+    // println!("radius: {:?}", radius);
+
+}
